requirements.txt
#pip install -r requirements.txt
django-recaptch
--upgrade django-crispy-forms
django-ckeditor
